update search results page
