+++ components/AddToGroup/Container/AddToGroupContainer.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import axios from "axios";
 import { useSession } from "next-auth/react";
 import { useState } from "react";
